Tweaked libtool LDFLAGS so that DLLs are built on Windows.  The trick is to add "-no-undefined" -- see issue37.
In order for this to work, the external package 'proj' must be tweaked
in a similar fashion; i.e., "-no-undefined" must be added to
libproj_la_LDFLAGS (happily, they use Libtool as well).  Libtool
refuses to link a DLL against a static Libtool library.  This is
prudent, but in this case it would probably work (I think
'spatialdata' is the only code that calls 'proj' directly).

For the Python extension modules, I also added "-avoid-version" so
that we arrive at "geocoordsmodule.dll" instead of
"geocoordsmodule-0.dll".